ResultsClient -> ResultClient
Also split up externalstorage components into their own files
