+++ gufe/storage/externalresource/__init__.py
@@ -1,0 +1,3 @@
+from .base import ExternalStorage
+from .filestorage import FileStorage
+from .memorystorage import MemoryStorage
